fix for bug [ 948279 ] upgrade postgresql 0.9.14-0.9.15 - there was a direct sql statement to drop an index, which is mysql only ==> put an if around it, to get rid of the pgSql error
git-svn-id: http://svn.egroupware.org/egroupware/trunk@15644 923e3c2e-ae13-0410-ba18-ced7211e33de

+++ projects/setup/tables_update.inc.php
@@ -639,7 +639,12 @@
 	$test[] = '0.8.7.019';
 	function projects_upgrade0_8_7_019()
 	{
-		$GLOBALS['phpgw_setup']->oProc->query('alter table phpgw_p_projects drop INDEX p_number');
+		// as the following SQL (which should'nt be here anyway) is mysql only, 
+		// I put an if around it, to fix other db's errors -- RalfBecker 2004/07/07
+		if ($GLOBALS['phpgw_setup']->oProc->sType == 'mysql')
+		{
+			$GLOBALS['phpgw_setup']->oProc->query('alter table phpgw_p_projects drop INDEX p_number');
+		}
 		$GLOBALS['phpgw_setup']->oProc->RenameColumn('phpgw_p_projects','id','project_id');
 		$GLOBALS['setup_info']['projects']['currentver'] = '0.8.7.020';
 		return $GLOBALS['setup_info']['projects']['currentver'];
